requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
openai>=1.10.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
pandas>=2.2.0
pillow>=10.3.0
PyMuPDF>=1.23.26
pdf2image>=1.17.0
python-dotenv>=1.0.0
aiofiles>=23.2.1

